+++ Utf8.cpp
@@ -17,7 +17,7 @@ void Utf8::reserve(int newSize)
     auto lastAllocated = allocated;
     allocated *= 2;
     allocated      = max(allocated, newSize);
-    allocated      = g_Allocator.alignSize(allocated);
+    allocated      = (int) g_Allocator.alignSize(allocated);
     auto newBuffer = (char*) g_Allocator.alloc(allocated);
     if (count)
         Memcpy(newBuffer, buffer, count + 1);
